Somascan panel

This is based on panel used in Sun et al. (2018) .

Usage

SomaScan160410

Format

A data frame with 5178 rows and 10 variables:

SOMAMER_ID

Somamer id

UniProt

UniProt id

Target

Protein target

TargetFullName

Protein target full name

chr

chromosome (1-22,X,Y)

start

start

end

end

entGene

entrez gene

ensGene

ENSEMBL gene

extGene

external gene

Details

from the INTERVAL study.

References

Sun BB, Maranville JC, Peters JE, Stacey D, Staley JR, Blackshaw J, Burgess S, Jiang T, Paige E, Surendran P, Oliver-Williams C, Kamat MA, Prins BP, Wilcox SK, Zimmerman ES, Chi A, Bansal N, Spain SL, Wood AM, Morrell NW, Bradley JR, Janjic N, Roberts DJ, Ouwehand WH, Todd JA, Soranzo N, Suhre K, Paul DS, Fox CS, Plenge RM, Danesh J, Runz H, Butterworth AS (2018). “Genomic atlas of the human plasma proteome.” Nature, 558(7708), 73-79. ISSN 1476-4687 (Electronic) 0028-0836 (Linking), doi:10.1038/s41586-018-0175-2 .