Supplementary table 3

Supplementary information for Zhao et al. (2023) .

Usage

scallop_inf1

Format

A data frame with 180 rows and 19 variables:

UniProt

UnitProt ID

Protein

Protein name

Protein_gene_symbol

Gene symbol

Chromosome

Chromosome

Position

Position

cistrans

cis/trans

rsid

reference sequence ID

Effect_allele

Effect allele

Other_allele

Eeference allele

EAF

Effect allele frequency

b

b

SE

SE

log10P

log10(P)

Direction

Direction field in METAL output

HetISq

$I^2$

HetChiSq

Heterogeneity chi-square

HetDf

degrees of freedom

logHetP

Heterogeneity log10(P)

N

N

References

Zhao JH, Stacey D, Eriksson N, Macdonald-Dunlop E, Hedman ÅK, Kalnapenkis A, Enroth S, Cozzetto D, Digby-Bell J, Marten J, Folkersen L, Herder C, Jonsson L, Bergen SE, Gieger C, Needham EJ, Surendran P, Team EBR, Paul DS, Polasek O, Thorand B, Grallert H, Roden M, Võsa U, Esko T, Hayward C, Johansson Å, Gyllensten U, Powell N, Hansson O, Mattsson-Carlgren N, Joshi PK, Danesh J, Padyukov L, Klareskog L, Landén M, Wilson JF, Siegbahn A, Wallentin L, Mälarstig A, Butterworth AS, Peters JE (2023). “Genetics of circulating inflammatory proteins identifies drivers of immune-mediated disease risk and therapeutic targets.” Nature Immunology, 24(9), 1540-1551. doi:10.1038/s41590-023-01588-w .